+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: enogaWa <enogawa@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/04 19:45:01 by enogaWa           #+#    #+#             */
-/*   Updated: 2023/05/05 13:41:43 by enogaWa          ###   ########.fr       */
+/*   Updated: 2023/05/06 16:41:15 by enogaWa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,35 +20,6 @@
 #include <readline/history.h>
 
 t_return_error	g_return_error;
-
-void	free_token(t_tk **token)
-{
-	t_tk	*tmp;
-
-	if (!token)
-		return ;
-	while (*token)
-	{
-		tmp = (*token)->next;
-		free((*token)->word);
-		free(*token);
-		*token = tmp;
-	}
-}
-
-void	free_node(t_node **node)
-{
-	if (!node)
-		return ;
-	if ((*node)->redirect)
-		free_node(&(*node)->redirect);
-	if ((*node)->pipe)
-		free_node(&(*node)->pipe);
-	if ((*node)->token)
-		free_token(&(*node)->token);
-	free((*node)->filename);
-	free(*node);
-}
 
 void	init_return_error(void)
 {
@@ -75,10 +46,25 @@ void	interpret(char *line, t_env **env_list)
 	if (g_return_error.parse_error)
 		return ;
 	exec_cmd(node, env_list);
-//	print_t_tk(token);
-//	print_node(node, 0);
 	free_token(&token);
 	free_node(&node);
+}
+
+int	into_minishell(char *line, t_env *env_list)
+{
+	if (!line)
+	{
+		if (isatty(STDIN_FILENO))
+			wrap_write (STDOUT_FILENO, "exit\n", strlen("exit\n"));
+		return (1);
+	}
+	if (*line)
+	{
+		add_history(line);
+		interpret(line, &env_list);
+	}
+	free(line);
+	return (0);
 }
 
 int	main(int argc, char const *argv[], char *envp[])
@@ -95,23 +81,8 @@ int	main(int argc, char const *argv[], char *envp[])
 	{
 		init_return_error();
 		line = readline(PROMPT);
-		if (!line)
-		{
-//			wrap_write (STDOUT_FILENO, "exit", strlen("exit\n"));
-			if (isatty(STDIN_FILENO))
-			{
-				wrap_write (STDOUT_FILENO, "exit", strlen("exit"));
-				if (g_return_error.ctrl_c == false)
-					wrap_write(STDOUT_FILENO, "\n", 1);
-			}
+		if (into_minishell(line, env_list))
 			break ;
-		}
-		if (*line)
-		{
-			add_history(line);
-			interpret(line, &env_list);
-		}
-		free(line);
 	}
 	free_env(&env_list);
 	return (g_return_error.return_value);
