+++ src/pipe/pipe.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pipe.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yshimoda <yshimoda@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: enogaWa <enogawa@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/12 14:43:45 by yshimoda          #+#    #+#             */
-/*   Updated: 2023/03/15 12:17:49 by yshimoda         ###   ########.fr       */
+/*   Updated: 2023/05/06 20:33:01 by enogaWa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,26 +30,41 @@ void	input_pipefd(t_node *node, int *inpipe)
 	}
 }
 
-bool	is_redirect_in_exist(t_node *node)
-{
-	while (node)
-	{
-		if (node->kind == ND_REDIRECT_IN || node->kind == ND_REDIRECT_HEREDOC)
-			return (true);
-		node = node->redirect;
-	}
-	return (false);
-}
+// bool	is_redirect_in_exist(t_node *node)
+// {
+// 	while (node)
+// 	{
+// 		if (node->kind == ND_REDIRECT_IN || node->kind == ND_REDIRECT_HEREDOC)
+// 			return (true);
+// 		node = node->redirect;
+// 	}
+// 	return (false);
+// }
 
-bool	is_redirect_out_exist(t_node *node)
+// bool	is_redirect_out_exist(t_node *node)
+// {
+// 	while (node)
+// 	{
+// 		if (node->kind == ND_REDIRECT_OUT || node->kind == ND_REDIRECT_APPEND)
+// 			return (true);
+// 		node = node->redirect;
+// 	}
+// 	return (false);
+// }
+
+void	connect_inpipe(t_node *node)
 {
-	while (node)
+	wrap_close(node->inpipe[1]);
+	if (!is_redirect_in_exist(node))
 	{
-		if (node->kind == ND_REDIRECT_OUT || node->kind == ND_REDIRECT_APPEND)
-			return (true);
-		node = node->redirect;
+		if (node->inpipe[0] != 0)
+		{
+			wrap_dup2(node->inpipe[0], STDIN_FILENO);
+			wrap_close(node->inpipe[0]);
+		}
 	}
-	return (false);
+	else
+		wrap_close(node->inpipe[0]);
 }
 
 void	connect_pipe(t_node *node)
@@ -71,38 +86,19 @@ void	connect_pipe(t_node *node)
 			wrap_close(node->outpipe[1]);
 	}
 	if (node->inpipe[0] != INT_MAX)
-	{
-		wrap_close(node->inpipe[1]);
-		if (!is_redirect_in_exist(node))
-		{
-			if (node->inpipe[0] != 0)
-			{
-				wrap_dup2(node->inpipe[0], STDIN_FILENO);
-				wrap_close(node->inpipe[0]);
-			}
-		}
-		else
-			wrap_close(node->inpipe[0]);
-	}
+		connect_inpipe(node);
 }
 
 void	reset_pipe_builtin(t_node *node)
 {
 	if (node->kind != ND_SIMPLE_CMD)
 		return ;
-//	if (node->fd_save_inpipe != INT_MAX)
-//	{
-//		wrap_close(STDIN_FILENO);
-//		wrap_dup2(node->fd_save_inpipe, STDIN_FILENO);
-//		wrap_close(node->fd_save_inpipe);
-//	}
 	if (node->fd_save_outpipe != INT_MAX)
 	{
 		wrap_close(STDOUT_FILENO);
 		wrap_dup2(node->fd_save_outpipe, STDOUT_FILENO);
 		wrap_close(node->fd_save_outpipe);
 	}
-
 }
 
 void	connect_pipe_builtin(t_node *node)
@@ -111,7 +107,6 @@ void	connect_pipe_builtin(t_node *node)
 		return ;
 	if (node->outpipe[1] != INT_MAX)
 	{
-		//wrap_close(node->outpipe[0]);
 		if (!is_redirect_out_exist(node))
 		{
 			if (node->outpipe[1] != 1)
@@ -119,26 +114,9 @@ void	connect_pipe_builtin(t_node *node)
 				node->fd_save_outpipe = wrap_dup(STDOUT_FILENO);
 				wrap_close(STDOUT_FILENO);
 				wrap_dup2(node->outpipe[1], STDOUT_FILENO);
-				//wrap_close(node->outpipe[1]);
 			}
 		}
 		else
 			wrap_close(node->outpipe[1]);
 	}
-//	if (node->inpipe[0] != INT_MAX)
-//	{
-//		wrap_close(node->inpipe[1]);
-//		if (!is_redirect_in_exist(node))
-//		{
-//			if (node->inpipe[0] != 0)
-//			{
-//				node->fd_save_inpipe = wrap_dup(STDIN_FILENO);
-//				wrap_close(STDIN_FILENO);
-//				wrap_dup2(node->inpipe[0], STDIN_FILENO);
-//				wrap_close(node->inpipe[0]);
-//			}
-//		}
-//		else
-//			wrap_close(node->inpipe[0]);
-//	}
 }
